Update schema_validator.py
Added a function that generates an iterator to retrieve all the IFC types (dicts in this case) from the evaluated files. For each IFC instance, it was checked if their parameters existed in the original schema.

This way, typing errors can be verified and the whole file can be validated. Originally this did not occur, since a wrong declared parameter was ignored by the jsonschema.validate function.